apps/api/pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "jarvis-api"
version = "0.4.0"
description = "JARVIS API — Phase 5 screen intelligence"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1",
  "uvicorn[standard]>=0.30.0,<1",
  "pydantic-settings>=2.5.0,<3",
  "httpx>=0.27.0,<1",
  "anyio>=4.0,<5",
  "aiosqlite>=0.20,<1",
  "crewai[tools]>=0.86.0,<1",
  "slack-sdk>=3.33.0,<4",
  "cryptography>=43.0.0,<46",
  "mss>=9.0.1,<11",
  "Pillow>=10.4.0,<12",
  "numpy>=1.26.0,<3",
  "opencv-python-headless>=4.10.0,<5",
  "pytesseract>=0.3.10,<1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0,<9",
  "pytest-asyncio>=0.24.0,<1",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
